Initial commit with actual code.
A very simple python script that outputs a desktop notification when the
command expressed in the arguments completes.

I intentionally set shell=True in subprocess so that long pipelines and
complex shell operations (loops, etc.) can be called simply. I do not
believe this to be a security vulnerability since the user would already
have access to a shell in the first place.